Use "video" as default mimetype
Currently, the mimetype is defaulted to one of the standard mimetypes in
ShakaPlayer. If the content mimetype is not available, and default
mimetype is passed to the demuxer, then the demuxer assumes that
mimetype is valid and tries to process the stream accordingly. This is
not correct and fails if the content is not of default mimetype.

To avoid this, use a generic default value for mimetype so that the
demuxer internally identifies the type.

[Problem]
Unable to play quicktime content without mimetype information.

[Test]
The test content with issue is playing.